backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
python-multipart==0.0.20
sqlalchemy==2.0.41
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
celery==5.5.3
redis==6.2.0
opencv-python-headless==4.12.0.88
numpy==2.2.6
faiss-cpu==1.12.0
torch==2.8.0
python-dotenv==1.1.1
